Finished my first basic project, feels good onto the next project

+++ note-taker-app/app/notes/DeleteNoteButton.tsx
@@ -8,6 +8,9 @@ export default function DeleteNoteButton({ noteId }: { noteId: string }) {
     const router = useRouter();
 
     async function handleDelete(){
+        const confirmed = confirm("Are you sure you want to delete note?");
+
+        if (!confirmed) return;
         try {
             await deleteNote(noteId);
             router.refresh();
